+++ examples/config.example.php
@@ -1,0 +1,11 @@
+<?php
+
+/**
+ * Config file containing the credentials to connect to the service.
+ */
+
+// The service endpoint.
+$apiEndpoint = 'https://basisregisters.vlaanderen.be/api/v1';
+
+// Optional api endpoint user key.
+$apiUserKey = null;